+++ vedo/__init__.py
@@ -1,195 +1,32 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-.. image:: https://user-images.githubusercontent.com/32848391/46815773-dc919500-cd7b-11e8-8e80-8b83f760a303.png
-
-A python module for scientific analysis of 3D objects and
-point clouds based on [VTK](https://www.vtk.org/) and [numpy](http://www.numpy.org/).
-
-# Documentation
-Use this page to search and inspect `vedo` sub-modules, methods and functions.
-These documentation pages are automatically generated
-by [pdoc](https://pdoc3.github.io/pdoc/) from the source files.
-
-## Quick start
-```bash
-pip install vedo
-```
-Then
-```python
-from vedo import Cone
-cone = Cone()      # Create a simple cone
-cone.show(axes=1)  # Show it (with axes)
-```
-.. image:: https://vedo.embl.es/images/feats/cone.png
-
-
-## Command Line Interface
-The library comes with a convenient Command Line Interface. Type for example in your terminal:
-
-```bash
-vedo --help
-vedo https://vedo.embl.es/examples/data/panther.stl.gz
-```
-.. image:: https://vedo.embl.es/images/feats/vedo_cli_panther.png
-
-Pressing `h` will then show a number of options to interact with your 3D scene:
-```
- ==========================================================
-| Press: i     print info about selected object            |
-|        I     print the RGB color under the mouse         |
-|        <-->  use arrows to reduce/increase opacity       |
-|        w/s   toggle wireframe/surface style              |
-|        p/P   change point size of vertices               |
-|        l     toggle edges visibility                     |
-|        x     toggle mesh visibility                      |
-|        X     invoke a cutter widget tool                 |
-|        1-3   change mesh color                           |
-|        4     use data array as colors, if present        |
-|        5-6   change background color(s)                  |
-|        09+-  (on keypad) or +/- to cycle axes style      |
-|        k     cycle available lighting styles             |
-|        K     cycle available shading styles              |
-|        A     toggle anti-aliasing                        |
-|        D     toggle depth-peeling (for transparencies)   |
-|        o/O   add/remove light to scene and rotate it     |
-|        n     show surface mesh normals                   |
-|        a     toggle interaction to Actor Mode            |
-|        j     toggle interaction to Joystick Mode         |
-|        u     toggle perspective/parallel projection      |
-|        r     reset camera position                       |
-|        C     print current camera settings               |
-|        S     save a screenshot                           |
-|        E     export rendering window to numpy file       |
-|        q     return control to python script             |
-|        Esc   abort execution and exit python kernel      |
-|----------------------------------------------------------|
-| Mouse: Left-click    rotate scene / pick actors          |
-|        Middle-click  pan scene                           |
-|        Right-click   zoom scene in or out                |
-|        Cntrl-click   rotate scene                        |
- ==========================================================
-```
-
-### file format conversion
-You can convert on the fly a file (or multiple files) to a different format with
-```bash
-vedo --convert bunny.obj --to ply
-```
-
-### some useful bash aliases
-```bash
-alias vr='vedo --run '        # to search and run examples by name
-alias vs='vedo -i --search '  # to search for a string in examples
-alias ve='vedo --eog '        # to view single and multiple images
-alias vv='vedo -bg blackboard -bg2 gray3 -z 1.05 -k glossy -c blue9'
-```
-
-## Example Galleries
-Check out the example galleries organized by subject [**here**](https://vedo.embl.es/#gallery).
-
-
-## Running on a headless server
-- Install `libgl1-mesa` and `xvfb` on your server:
-```bash
-sudo apt install libgl1-mesa-glx libgl1-mesa-dev xvfb
-pip install vedo
-```
-
-- Execute on startup:
-```bash
-set -x
-export DISPLAY=:99.0
-which Xvfb
-Xvfb :99 -screen 0 1024x768x24 > /dev/null 2>&1 &
-sleep 3
-set +x
-exec "$@"
-```
-
-- You can save the above code above as `/etc/rc.local` and use `chmod +x` to make it executable.
-    It may throw an error during startup. Then test it with, e.g.:
-```python
-import vedo
-plt = vedo.Plotter(offscreen=True, size=(500,500))
-plt.show(vedo.Cube()).screenshot('mycube.png').close()
-```
-
-## Running in a Docker container
-You need to set everything up for offscreen rendering: there are two main ingredients
-
-- `vedo` should be set to render in offscreen mode
-- guest OS in the docker container needs the relevant libraries installed
-    (in this example we need the Mesa openGL and GLX extensions, and Xvfb to act as a virtual screen.
-    It's maybe also possible to use OSMesa offscreen driver directly, but that requires a custom
-    build of VTK).
-
-- Create a `Dockerfile`:
-```bash
-FROM python:3.8-slim-bullseye
-
-RUN apt-get update -y \
-  && apt-get install libgl1-mesa-dev libgl1-mesa-glx xvfb -y --no-install-recommends \
-  && apt-get purge -y --auto-remove -o APT::AutoRemove::RecommendsImportant=false \
-  && rm -rf /var/lib/apt/lists/*
-RUN pip install vedo && rm -rf $(pip cache dir)
-RUN mkdir -p /app/data
-
-WORKDIR /app/
-COPY test.py set_xvfb.sh /app/
-ENTRYPOINT ["/app/set_xvfb.sh"]
-```
-
-- `set_xvfb.sh`:
-```bash
-#!/bin/bash
-set -x
-export DISPLAY=:99.0
-Xvfb :99 -screen 0 1024x768x24 > /dev/null 2>&1 &
-#sleep 3
-set +x
-exec "$@"
-```
-
-- `test.py`:
-```python
-from vedo import Sphere, Plotter, settings
-settings.screenshotTransparentBackground = True
-sph = Sphere(pos=[-5, 0, 0], c="r")
-plt = Plotter(interactive=False, offscreen=True)
-plt.show(sph)
-plt.screenshot("./data/out.png", scale=2).close()
-```
-
-Then you can
-
-1. `$ docker build -t vedo-test-local .`
-2. `$ docker run --rm -v /some/path/output:/app/data vedo-test-local python test.py` (directory `/some/path/output` needs to exist)
-3. There should be an `out.png` file in the output directory.
-
-
-## Getting help
-Check out the [**Github repository**](https://github.com/marcomusy/vedo)
-for more information, where you can ask questions and report issues.
-You are also welcome to post specific questions on the [**image.sc**](https://forum.image.sc/) forum.
+.. include:: ../docs/documentation.md
 """
 ##### To generate documentation #######################################################
 # cd Projects/vedo
+# pip uninstall vedo
 # pdoc --html . --force -c lunr_search="{'fuzziness': 0, 'index_docstrings': True}"
-######################################################################### pdoc excludes
+# chmod 755 html/ -R
+# mount_staging
+# rm ~/Projects/StagingServer/var/www/html/vtkplotter.embl.es/autodocs/html
+# mv html/ ~/Projects/StagingServer/var/www/html/vtkplotter.embl.es/autodocs/
+############## pdoc excludes
 __pdoc__ = {}
 __pdoc__['embedWindow'] = False
 __pdoc__['backends'] = False
 __pdoc__['cli'] = False
 __pdoc__['cmaps'] = False
 __pdoc__['version'] = False
+__pdoc__['base.BaseActor.getPointArray'] = False
+__pdoc__['base.BaseActor.getCellArray'] = False
+__pdoc__['base.BaseActor.addPointArray'] = False
+__pdoc__['base.BaseActor.addCellArray'] = False
 __pdoc__['pointcloud.Points.pointColors'] = False
 __pdoc__['pointcloud.Points.cellColors'] = False
 __pdoc__['pointcloud.Points.thinPlateSpline'] = False
 __pdoc__['pointcloud.Points.warpByVectors'] = False
 __pdoc__['pointcloud.Points.distanceToMesh'] = False
-__pdoc__['dolfin.show'] = False
-__pdoc__['pyplot.show'] = False
 
 
 #######################################################################################